Raise warning to ensure early notification
I suggest raising the warning minimum score a lot (0.95 in this example) to ensure SNO's actually get an early warning. Any score dropping below 0.95 is already a good indication something is wrong. If you want to be conservative I can imagine going with 0.90, but 0.76 means you're already 2/3rds of the way towards serious consequences for the node. This is especially important since the suspension threshold for down time is likely to go up in the future as well.